+++ docker-compose.yaml
@@ -27,7 +27,7 @@ services:
   ssh:
     build:
       context: .
-      target: mutagen-ssh
+      target: ssh
     ports:
       - 2222:2222
       - 6061:6060
@@ -36,8 +36,8 @@ services:
       - --ssh-listen-addr=0.0.0.0:2222
       - --ssh-key-path=/keys/id_ed25519
     volumes:
-      - ./api/tmp/repos:/repos
-      - ./mutagen-ssh/cmd/mutagen-ssh/keys:/keys
+      - ./tmp/repos:/repos
+      - ./ssh/cmd/ssh/keys:/keys
 
 volumes:
   db-data:
