+++ ssh/cmd/ssh/generate-development.sh
